software/data-analysis/requirements.txt
# Smart Posture Tracker - Python Dependencies
# Install with: pip install -r requirements.txt

# Core data analysis libraries
pandas>=2.0.0
numpy>=1.24.0

# Serial communication
pyserial>=3.5

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Scientific computing
scipy>=1.10.0

# Machine learning (for future posture classification)
scikit-learn>=1.3.0

# Real-time plotting (optional)
plotly>=5.15.0

# Development tools
jupyter>=1.0.0
ipython>=8.0.0